refactor: collected tests in guessTheFunction_test.go, removed isExpr() from expr
